refactor: analytics indexing dto
